Split components into different files - Split all components into different files - Limited the API call rate
