+++ public/src/index.css
@@ -3,6 +3,7 @@ body {
   background-color: #242f39;
   padding: 0;
   background-image: url("images/background.jpg");
+  overflow-x: hidden;
 }
 @font-face {
     font-family: 'Radiance';
@@ -37,90 +38,6 @@ body {
     font-style: normal;
 }
 
-
-h1 {
-  font-family:'Radiance-Bold', sans-serif;
-  font-size: 5vw;
-  color: #af3d27;
-   display: inline;
-}
-
-h2 {
-  font-family:'Radiance-Bold', sans-serif;
-  font-size: 2vw;
-  color: #af3d27;
-}
-
-h3 {
-  font-family:'Radiance-SemiBold', sans-serif;
-  font-size: 2.88vw;
-  margin-top:0vh;
-  color: #336391;
-}
-
-#positivityScore {
-  text-align: center;
-  width: 100%;
-  height: 100%;
-  float: left;
-  background-color: transparent;
-
-}
-
-#negativityScore {
-  text-align: center;
-  width: 50%;
-  height: 100%;
-  float: right;
-  background-color: #9c4345;
-}
-
-i {
-  position:relative;
-  margin-top: 40vh;
-  border: solid white;
-  border-width: 0 1.3vw 1.3vw 0;
-  padding: 1.3vw;
-}
-
-.arrow-right {
-  float: right;
-  margin-right: 5vw;
-    transform: rotate(-45deg);
-    -webkit-transform: rotate(-45deg);
-    z-index: 10;
-
-}
-
-.arrow-left {
-  float:left;
-    margin-left: 5vw;
-    transform: rotate(135deg);
-    -webkit-transform: rotate(135deg);
-    z-index: 10;
-
-}
-
-.fade-in{
-  -webkit-transition: 0.5s;
-  -moz-transition: 0.5s;
-  -o-transition: 0.5s;
-  transition: 1.5s;
-  opacity: 0;
-  visibility: hidden;
-}
-
-.fade-in.show{
-  opacity: 1;
-  visibility: visible;
-}
-.center {
-   width: 100vw;
-   position: absolute;
-   left: 50%;
-   margin-left: -50vw;
-   margin-top: 35vh;
-}
 .myButton{
     background:url("images/loginButton.png");
     background-size: 100% 100%;
@@ -128,36 +45,4 @@ i {
     border:0px;
     width:171px;
     height: 65px;
-}
-
-.sentence {
-  color: #000000;
-}
-
-.example-enter {
-    opacity: 0;
-    transform:   translate(-250px,0);
-    transform: translate3d(-250px,0,0);
-}
-.example-enter.example-enter-active {
-	opacity: 1;
-	transition: opacity 1s ease;
-	transform:   translate(0,0);
-	transform: translate3d(0,0,0);
-	transition-property: transform, opacity;
-	transition-duration: 300ms;
-	transition-timing-function: cubic-bezier(0.175, 0.665, 0.320, 1), linear;
-}
-.example-leave {
-	opacity: 1;
-	transform:   translate(0,0,0);
-	transform: translate3d(0,0,0);
-  transition-property: transform, opacity;
-	transition-duration: 300ms;
-	transition-timing-function: cubic-bezier(0.175, 0.665, 0.320, 1), linear;
-}
-.example-leave.example-leave-active {
-	opacity: 0;
-	transform:   translate(250px,0);
-	transform: translate3d(250px,0,0);
 }
